requirements.txt
\
streamlit==1.37.1
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.4.0
transformers==4.44.2
torch>=2.2.0
sentencepiece>=0.2.0
